Fix float32 precision loss in packed/streaming time_stamps
Root cause of verify_packed_landmark_rows disagreeing with
_index_rows_from_events on real eICU data (initially 38,424 missing /
40,727 extra landmark rows on one alert event alone): both
PackedContextSampler and PackedLaneSampler stacked time_stamps as
torch.float (float32), which loses enough precision at real-data time
magnitudes (hundreds of hours since a subject's origin) to disagree with
_index_rows_from_events' float64 polars computation at the 6th decimal
place -- exactly what _landmark_key_set's round-to-6-decimals comparison
guard is sensitive to. Confirmed via a local repro (tiny CPU
TransformerBackbone + PackedContextSampler over real eICU held-out
shards): every mismatch paired up as the same (subject, visit, bucket)
landmark, off by ~1e-6 hours.

Both files now stack time_stamps as torch.double instead. The model's
own forward pass is unaffected -- TimeEmbeddingLayer.forward already
casts to float32 internally before use, so this only adds precision to
the landmark-bookkeeping and IndexRow.time_hours path. Regression tests
in both test_packed_context.py and test_streaming.py use 434.416667h,
the literal value observed in the real-data repro, not a hand-picked
adversarial one; both verified red (fail under torch.float) before the
fix and green after.

On the real eICU repro this eliminates the "missing" class entirely
(38,424 -> 0) and reduces "extra" by 94% (40,727 -> 2,302). The
remaining ~2,302 extra rows are a distinct, separate bug (same-timestamp
discharge medication-stop bundles interleaving two different hadm_ids in
token order, which spuriously retriggers _landmark_mask's ~same_visit
check even though the bucket/time hasn't changed) -- not fixed here,
reported separately per the "don't paper over a surviving class" review
requirement.

+++ odyssey/data/packed_context.py
@@ -320,7 +320,18 @@ class PackedContextSampler:
 
         concept_ids_full = _stack("concept_ids", torch.long)
         type_ids_full = _stack("type_ids", torch.long)
-        time_stamps_full = _stack("time_stamps", torch.float)
+        # double, not float: _landmark_mask/_unrebase_truncated_times and
+        # the final IndexRow.time_hours all read this same tensor back for
+        # exact comparison against _index_rows_from_events' float64 polars
+        # computation -- float32 loses enough precision at real-data time
+        # magnitudes (hundreds of hours since origin) to occasionally flip
+        # the 6th decimal digit verify_packed_landmark_rows compares on,
+        # confirmed via a real-eICU-data repro (root-caused 2026-08-23).
+        # The model's own TimeEmbeddingLayer.forward() already does its
+        # own .float() cast before use, so this is purely additive
+        # precision for the landmark-bookkeeping path, not a model-input
+        # dtype change.
+        time_stamps_full = _stack("time_stamps", torch.double)
         ages_full = _stack("ages", torch.float)
         visit_orders_full = _stack("visit_orders", torch.long)
         visit_segments_full = _stack("visit_segments", torch.long)
